refactor: create own class for request objects an contexts
