+++ packages/ui5-cc-spreadsheetimporter/src/i18n/i18n_en.properties
@@ -103,7 +103,7 @@ spreadsheetimporter.busyDialogTitle = Upload in progress ...
 
 # Confirm Dialog for Continue on Not Found Object in Update Operation
 spreadsheetimporter.confirmTitle=Confirmation Required
-spreadsheetimporter.confirmMessage=Do you want to continue uploading the found records?\n Not found objects will be ignored.
+spreadsheetimporter.confirmMessage=Do you want to continue uploading the found records?\n Not found objects will be ignored and false draft entities will be corrected.
 spreadsheetimporter.continue=Continue
 spreadsheetimporter.cancel=Cancel
